Allow multiple controllers
The `controller` array in config.h lists the controllers, their pin
assignments, and their keymaps.  Any number of controller definitions
may be added.

To do this efficiently, there is a `snes_read_multi` function which
reads from multiple controllers at once.  If all the controllers share
their CLOCK or LATCH pins, this function can be instructed to
clock/latch them simultaneously.  This is a pretty big win, since
polling consists mainly of busy-waiting.

It is no longer possible to specify the GPIO pins to use on the command
line.  This is unfortunate, but it's probably not a good idea to be
hotplugging over GPIO anyways, so a static configuration makes sense.  A
`--<pin name> <controller>:<pin number>` scheme could be used, but is
not implemented as of this commit.